authentication guards added to graphQL

+++ app/GraphQL/Mutations/PostMutation.php
@@ -3,6 +3,8 @@
 namespace App\GraphQL\Mutations;
 
 use App\Models\Post;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Log;
 
 final class PostMutation
 {
@@ -11,15 +13,26 @@ final class PostMutation
         $post = new Post();
         $post->title = $args['title'];
         $post->content = $args['content'];
-        $post->user_id = $args['user_id'];
+        $post->user_id = auth()->user()->id;
         $post->save();
 
-        return $post;
+        return [
+            'success' => true,
+            'message' => 'Post created successfully',
+            'data' => $post
+        ];
     }
 
     public function updatePost($root, array $args)
     {
         $post = Post::findOrFail($args['id']);
+        if($post->user_id != auth()->user()->id){
+            return [
+                'success' => false,
+                'message' => 'User not authorized to update this post'
+            ];
+        }
+        
         if (isset($args['title'])) {
             $post->title = $args['title'];
         }
@@ -28,15 +41,28 @@ final class PostMutation
         }
         $post->save();
 
-        return $post;
+        return [
+            'success' => true,
+            'message' => 'Post updated successfully',
+            'data' => $post
+        ];
     }
 
     public function deletePost($root, array $args)
     {
         $post = Post::findOrFail($args['id']);
+        if($post->user_id != auth()->user()->id){
+            return [
+                'success' => false,
+                'message' => 'User not authorized to delete this post'
+            ];
+        }
         $post->delete();
 
-        return $post;
+        return [
+            'success' => true,
+            'message' => 'Post deleted successfully'
+        ];
     }
 
 }
